+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+## [1.520.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.520.0...v1.520.1) (2026-08-12)
+
+
+### Bug Fixes
+
+* **cdn-analysis:** exclude Adobe internal user agents from agentic traffic (LLMO-6961) ([#2887](https://github.com/adobe/spacecat-audit-worker/issues/2887)) ([8aecb75](https://github.com/adobe/spacecat-audit-worker/commit/8aecb75605ed68ee66ef50c72e2a3ec904cfd142))
+
+# [1.520.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.519.0...v1.520.0) (2026-08-12)
+
+
+### Features
+
+* **offsite:** structured key=value logging for offsite-brand-presence ([#2837](https://github.com/adobe/spacecat-audit-worker/issues/2837)) ([61e9763](https://github.com/adobe/spacecat-audit-worker/commit/61e9763b06f8ef1725265a808ff79f254254aa82))
+
+# [1.519.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.518.3...v1.519.0) (2026-08-12)
+
+
+### Features
+
+* **toc:** freeze suggestions with data.isVerified=true from re-audit… ([#2885](https://github.com/adobe/spacecat-audit-worker/issues/2885)) ([19ac0d0](https://github.com/adobe/spacecat-audit-worker/commit/19ac0d0136eae22191ef63a5a0092b66966ce809))
+
+## [1.518.3](https://github.com/adobe/spacecat-audit-worker/compare/v1.518.2...v1.518.3) (2026-08-12)
+
+
+### Bug Fixes
+
+* **audits:** query RUM by hostname for sub-path sites (cwv, internal-links, forms, alt-text) ([#2881](https://github.com/adobe/spacecat-audit-worker/issues/2881)) ([11236a5](https://github.com/adobe/spacecat-audit-worker/commit/11236a54b9d641b08e3aa75dac8e83e5b08e5df2))
+* **backlinks:** scope broken-backlink results to the site sub-path ([#2882](https://github.com/adobe/spacecat-audit-worker/issues/2882)) ([070abe5](https://github.com/adobe/spacecat-audit-worker/commit/070abe5f615ecbfd795bcdf2af7863096484d26e))
+* **metatags:** emit absolute suggestion/page URLs for sub-path sites (metatags + product-metatags) ([#2880](https://github.com/adobe/spacecat-audit-worker/issues/2880)) ([f663ea3](https://github.com/adobe/spacecat-audit-worker/commit/f663ea36f3d5f6211db051acc7d6799a7deec460))
+
+## [1.518.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.518.1...v1.518.2) (2026-08-11)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#2879](https://github.com/adobe/spacecat-audit-worker/issues/2879)) ([151c5e6](https://github.com/adobe/spacecat-audit-worker/commit/151c5e64a9232019dcf8d9f68082e71b73e20d17))
+
 ## [1.518.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.518.0...v1.518.1) (2026-08-11)